backend/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-colorfield==0.7.2
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf_extra_fields==3.4.1
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
pytz==2022.2.1
reportlab==3.6.11
simplejwt==2.0.1
sqlparse==0.4.2
